+++ src/Script.hs
@@ -25,13 +25,13 @@ data Setup = Setup
   , noskel :: Bool
   , username :: T.Text
   , program :: String
-  , mhome :: Maybe String
+  , createhome :: Bool
   , homedir :: T.Text
   , mprojectDir :: Maybe FilePath
   }
 
-setupScript :: Bool -> Bool -> Setup -> String
-setupScript dbg haveRunuser (Setup {..}) =
+setupScript :: Bool -> Bool -> Bool -> Setup -> String
+setupScript dbg haveRunuser haveSudo (Setup {..}) =
   T.unpack . T.replace "\t" " " . linearScript $
   sudoSetup >> homeSetup
   where
@@ -41,7 +41,7 @@ setupScript dbg haveRunuser (Setup {..}) =
     -- haveCmd c = runHide "command" ["-v",c]
 
     sudoSetup =
-      when haveRunuser $
+      when haveSudo $
       if nosudo
       -- FIXME actual sudo path
       then runHide "rm" ["-f", "/usr/bin/sudo"]
@@ -52,22 +52,18 @@ setupScript dbg haveRunuser (Setup {..}) =
           runHide "chmod" ["440", T.pack sudoers]
 
     homeSetup = do
-      when haveRunuser $
-        when (isNothing mhome) $ do
+      when createhome $ do
         runHide "mkdir" ["-p", homedir]
         runHide "chown" [username, homedir]
-        -- FIXME? might be more consist not to condition on .bashrc
       unless noskel $
-          whenCmd
-          (test (TDirExists homedir)
-           -&&-
-           -- shell-monad bug: TAnd emits raw &&
-           test (TNot $ TFileExists (T.unpack homedir </> ".bashrc"))
-           -&&-
-           test (TDirExists (T.pack "/etc/skel"))) $
-          let cpArgs = ["-a", "--update=none", "/etc/skel/.", homedir <> "/"]
-          in if haveRunuser
-             then runHide "runuser" $ ["-u", username, "--"] ++ "cp" : cpArgs
-             else runHide "cp" cpArgs
+        when haveRunuser $
+        whenCmd
+        (test (TDirExists homedir)
+         -&&-
+         test (TDirExists (T.pack "/etc/skel"))) $
+        let cpArgs = ["-a", "--update=none", "/etc/skel/.", homedir <> "/"]
+        in if haveRunuser
+           then runHide "runuser" $ ["-u", username, "--"] ++ "cp" : cpArgs
+           else runHide "cp" cpArgs
       when (isNothing mprojectDir) $
-        runHide "cd" []
+        runHide "cd" $ if createhome then [] else [homedir]